Make Collection Modal Component and Create Collection from Add To Collection Dropdown
