+++ app/components/collections-flyout/component.js
@@ -1,4 +1,3 @@
-import { schedule } from '@ember/runloop';
 import { inject as service } from '@ember/service';
 import { get, computed } from '@ember/object';
 import Component from '@ember/component';
@@ -7,11 +6,9 @@ export default Component.extend({
   session: service(),
   store: service(),
   collectionToDelete: null,
-  errorMessage: null,
   showConfirmation: false,
   showDeleteButtons: false,
   showModal: false,
-
   collections: computed('store', {
     get() {
       if (!get(this, 'session.isAuthenticated') ||
@@ -24,28 +21,8 @@ export default Component.extend({
   }),
 
   actions: {
-    /**
-     * Action to create a new collection
-     */
-    addNewCollection() {
-      const name = this.get('name');
-      const description = this.get('description');
-
-      schedule('actions', () => {
-        const newCollection = this.get('store').createRecord('collection', {
-          name,
-          description
-        });
-
-        return newCollection.save()
-          .then(() => {
-            this.set('showModal', false);
-          })
-          .catch((error) => {
-            newCollection.destroyRecord();
-            this.set('errorMessage', error.errors[0].detail);
-          });
-      });
+    openModal() {
+      this.set('showModal', true);
     },
     /**
      * Action to cancel the deletion of a collection
@@ -81,14 +58,6 @@ export default Component.extend({
      * Action to open / close the create collection modal
      * @param {boolean} open - whether modal should be open
      */
-    setModal(open) {
-      if (!open) {
-        this.set('name', null);
-        this.set('description', null);
-        this.set('errorMessage', null);
-      }
-      this.set('showModal', open);
-    },
     toggleEdit() {
       this.set('showDeleteButtons', !this.get('showDeleteButtons'));
     }
