+++ tests/integration/components/collections-flyout/component-test.js
@@ -13,13 +13,6 @@ const mockCollection = {
   get: name => name
 };
 
-const collectionModel = {
-  save() {
-    return new EmberPromise(resolve => resolve(mockCollection));
-  },
-  destroyRecord() {}
-};
-
 moduleForComponent('collections-flyout', 'Integration | Component | collections flyout', {
   integration: true
 });
@@ -82,7 +75,14 @@ test('it opens collection create modal', function (assert) {
   this.set('collections', []);
   this.set('showModal', false);
 
-  this.render(hbs`{{collections-flyout collections=collections showModal=showModal}}`);
+  const setModal = () => { this.set('showModal', true); };
+
+  this.set('setModal', setModal);
+
+  this.render(hbs`{{collections-flyout
+    collections=collections
+    showModal=showModal
+    setModal=setModal}}`);
   assert.equal(this.get('showModal'), false);
   // Make sure there are no modals
   assert.notOk($('.modal').length);
@@ -103,54 +103,6 @@ test('it opens collection create modal', function (assert) {
     assert.equal(cancelButton.text().trim(), 'Cancel');
     assert.equal(createButton.text().trim(), 'Create');
   });
-});
-
-test('it creates a collection', function (assert) {
-  assert.expect(4);
-
-  injectSessionStub(this);
-
-  const $ = this.$;
-  const storeStub = EmberObject.extend({
-    createRecord(model, data) {
-      assert.strictEqual(model, 'collection');
-      assert.deepEqual(data, {
-        name: 'Test',
-        description: 'Test description'
-      });
-
-      return collectionModel;
-    },
-    findAll() {
-      return new EmberPromise(resolve => resolve([mockCollection]));
-    }
-  });
-
-  this.set('collections', []);
-  this.set('showModal', false);
-  this.set('name', null);
-  this.set('description', null);
-
-  this.register('service:store', storeStub);
-  this.inject.service('store');
-
-  this.render(hbs`{{collections-flyout
-    collections=collections
-    showModal=showModal
-    name=name
-    description=description
-  }}`);
-
-  $('.new').click();
-
-  this.set('name', 'Test');
-  this.set('description', 'Test description');
-
-  assert.ok(this.get('showModal'));
-
-  $('.collection-form__create').click();
-
-  assert.notOk(this.get('showModal'));
 });
 
 test('it renders an active collection', function (assert) {
@@ -219,22 +171,6 @@ test('it fails to create a collection', function (assert) {
   assert.ok(this.get('showModal'));
   assert.strictEqual($('.alert-warning > span').text().trim(),
     'This is an error message');
-});
-
-test('it cancels creation of a collection', function (assert) {
-  const $ = this.$;
-
-  injectSessionStub(this);
-
-  this.set('collections', []);
-  this.set('showModal', false);
-
-  this.render(hbs`{{collections-flyout collections=collections showModal=showModal}}`);
-
-  $('.new').click();
-  assert.ok(this.get('showModal'));
-  $('.collection-form__cancel').click();
-  assert.notOk(this.get('showModal'));
 });
 
 test('it deletes a collection', function (assert) {
